+++ src/main/java/api/CommonApiWrapper.java
@@ -1,16 +1,19 @@
 package api;
 
-import java.io.IOException;
-import java.util.logging.Logger;
-
 import org.apache.http.Header;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
-import org.apache.http.client.methods.*;
+import org.apache.http.client.methods.HttpDelete;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.client.methods.HttpPost;
+import org.apache.http.client.methods.HttpPut;
 import org.apache.http.entity.ContentType;
 import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.HttpClients;
 import org.apache.http.util.EntityUtils;
+
+import java.io.IOException;
+import java.util.logging.Logger;
 
 public class CommonApiWrapper {
     private static final Logger logger = Logger.getLogger(CommonApiWrapper.class.getName());
