+++ src/main/java/api/domain/product/Search.java
@@ -1,4 +1,4 @@
-package api.domain.catalog.product;
+package api.domain.product;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
