+++ src/main/java/api/domain/product/Variant.java
@@ -1,4 +1,4 @@
-package api.domain.catalog.product;
+package api.domain.product;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
